feat: create learning panel
Users who purchased a course can now learn the course by watching the videos and completing the lessons.
Users who have not purchased the course can preview the course by learning from the free chapters. When the user tries to access a paid chapter, a dialog will appear to prompt the user to purchase the course.
Multiple UI components were created to support the learning panel, such as the course header, chapters sidebar, lesson navigation, and complete lesson button.
The video player was refactored to be a shared component that can be reused in different parts of the application.
TODO: Add chessboard to lessons with chess puzzles.